Bugs fix

+++ AcademyProject/Controllers/AnswerController.cs
@@ -2,60 +2,67 @@
 using AcademyProject.Models;
 using AcademyProject.Services;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-
 namespace AcademyProject.Controllers
 {
-    [Route("api/[controller]")]
     [ApiController]
+    [Route("api/[controller]")]
+    [Authorize(Roles = "Administrators, Moderators, Lecturers, Students")]
     public class AnswerController : ControllerBase
     {
-        private readonly IGenericService<Answer> answerService;
+
         private readonly IMapper mapper;
-        public AnswerController(IGenericService<Answer> answerService, IMapper mapper)
+        private readonly SharedComponent component;
+        private readonly IGenericService<Answer> answerService;
+        private readonly IGenericService<User> userService;
+
+        public AnswerController(
+            IMapper mapper,
+            SharedComponent component,
+            IGenericService<Answer> answerService,
+            IGenericService<User> userService
+        )
         {
-            this.answerService = answerService;
             this.mapper = mapper;
-        }
-        // GET: api/<AnswerController>
-        [HttpGet]
-        public async Task<ActionResult<AnswerDTO>> Get()
-        {
-            var list = await answerService.GetAll();
-            var listAnswer = list.Select(x => mapper.Map<AnswerDTO>(x)).ToList();
-            return Ok(new { listAnswer });
+            this.component = component;
+            this.answerService = answerService;
+            this.userService = userService;
         }
 
-        // GET api/<AnswerController>/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<AnswerDTO>> Get(int id)
+        protected int GetCurrentUserId()
         {
-            var answer = await answerService.GetById(id);
-            if (answer == null)
-            {
-                return NotFound();
-            }
-            var answerDTO = mapper.Map<AnswerDTO>(answer);
-            return Ok(new { answerDTO });
+            string userId = User.Claims.Where(x => x.Type == "Id").FirstOrDefault()?.Value;
+            int id = Convert.ToInt32(userId);
+            return id;
         }
 
-        // POST api/<AnswerController>
+        // POST: api/<AnswerController>
         [HttpPost]
         public async Task<ActionResult<AnswerDTO>> Post([FromBody] AnswerDTO answerDTO)
         {
-            var answer = mapper.Map<Answer>(answerDTO);
-            answer = await answerService.Insert(answer);
-            answerDTO = mapper.Map<AnswerDTO>(answer);
-            return Ok(new { answer });
+            int userId = GetCurrentUserId();
+
+            Answer comment = new Answer();
+            comment.QuestionId = answerDTO.QuestionId;
+            comment.UserId = userId;
+            comment.Content = answerDTO.Content;
+            comment = await answerService.Insert(comment);
+            answerDTO = mapper.Map<AnswerDTO>(comment);
+
+            // Get comment user
+            var user = await userService.GetById(answerDTO.UserId);
+            answerDTO.User = mapper.Map<UserCommentDTO>(user);
+            answerDTO.User.Picture = await component.GetImageAsync(answerDTO.User.PictureId);
+            return Ok(answerDTO);
         }
 
-        // PUT api/<AnswerController>/5
+        // PUT: api/<AnswerController>/{id}
         [HttpPut("{id}")]
         public async Task<ActionResult<AnswerDTO>> Put(int id, [FromBody] AnswerDTO answerDTO)
         {
@@ -64,21 +71,36 @@ namespace AcademyProject.Controllers
             {
                 return NotFound();
             }
-            answer.QuestionId = answerDTO.QuestionId;
-            answer.UserId = answerDTO.UserId;
+
+            var userId = GetCurrentUserId();
+            if (userId != answer.UserId && !User.IsInRole("Administrators") && !User.IsInRole("Moderators"))
+            {
+                return Unauthorized();
+            }
+
             answer.Content = answerDTO.Content;
-            answer.CreatedAt = answerDTO.CreatedAt;
-            answer = await answerService.Update(answer);
+            answer.UpdatedAt = DateTime.Now;
+            await answerService.Update(answer);
 
-            answerDTO = mapper.Map<AnswerDTO>(answer);
-
-            return Ok(new { answerDTO });
+            answerDTO.UpdatedAt = answer.UpdatedAt;
+            return Ok(answerDTO);
         }
 
-        // DELETE api/<AnswerController>/5
+        // DELETE: api/<AnswerController>/{id}
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var answer = await answerService.GetById(id);
+            if (answer == null)
+            {
+                return NotFound();
+            }
+            int userId = GetCurrentUserId();
+            if (userId != answer.UserId && !User.IsInRole("Administrators") && !User.IsInRole("Moderators"))
+            {
+                return Unauthorized();
+            }
+
             await answerService.Delete(id);
             return Ok();
         }
