+++ AcademyProject/Controllers/BlogCommentController.cs
@@ -2,60 +2,88 @@
 using AcademyProject.Models;
 using AcademyProject.Services;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-
 namespace AcademyProject.Controllers
 {
-    [Route("api/[controller]")]
     [ApiController]
+    [Route("api/[controller]")]
+    [Authorize(Roles = "Administrators, Moderators, Lecturers, Students")]
     public class BlogCommentController : ControllerBase
     {
-        private readonly IGenericService<BlogComment> blogCommentService;
         private readonly IMapper mapper;
-        public BlogCommentController(IGenericService<BlogComment> blogCommentService, IMapper mapper)
+        private readonly SharedComponent component;
+        private readonly IGenericService<BlogComment> blogCommentService;
+        private readonly IGenericService<User> userService;
+
+        public BlogCommentController(
+            IMapper mapper,
+            SharedComponent component,
+            IGenericService<BlogComment> blogCommentService,
+            IGenericService<User> userService
+        )
         {
-            this.blogCommentService = blogCommentService;
             this.mapper = mapper;
+            this.component = component;
+            this.blogCommentService = blogCommentService;
+            this.userService = userService;
         }
+
+        protected int GetCurrentUserId()
+        {
+            string userId = User.Claims.Where(x => x.Type == "Id").FirstOrDefault()?.Value;
+            int id = Convert.ToInt32(userId);
+            return id;
+        }
+
         // GET: api/<BlogCommentController>
-        [HttpGet]
-        public async Task<ActionResult<BlogCommentDTO>> Get()
-        {
-            var list = await blogCommentService.GetAll();
-            var listBlogComment = list.Select(x => mapper.Map<BlogCommentDTO>(x)).ToList();
-            return Ok(new { listBlogComment });
-        }
+        //[HttpGet]
+        //public async Task<ActionResult<BlogCommentDTO>> Get()
+        //{
+        //    var list = await blogCommentService.GetAll();
+        //    var listBlogComment = list.Select(x => mapper.Map<BlogCommentDTO>(x)).ToList();
+        //    return Ok(listBlogComment);
+        //}
 
-        // GET api/<BlogCommentController>/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<BlogCommentDTO>> Get(int id)
-        {
-            var blogComment = await blogCommentService.GetById(id);
-            if (blogComment == null)
-            {
-                return NotFound();
-            }
-            var blogCommentDTO = mapper.Map<BlogCommentDTO>(blogComment);
-            return Ok(new { blogCommentDTO });
-        }
+        // GET: api/<BlogCommentController>/{id}
+        //[HttpGet("{id}")]
+        //public async Task<ActionResult<BlogCommentDTO>> Get(int id)
+        //{
+        //    var blogComment = await blogCommentService.GetById(id);
+        //    if (blogComment == null)
+        //    {
+        //        return NotFound();
+        //    }
+        //    var blogCommentDTO = mapper.Map<BlogCommentDTO>(blogComment);
+        //    return Ok(blogCommentDTO);
+        //}
 
-        // POST api/<BlogCommentController>
+        // POST: api/<BlogCommentController>
         [HttpPost]
         public async Task<ActionResult<BlogCommentDTO>> Post([FromBody] BlogCommentDTO blogCommentDTO)
         {
-            var blogComment = mapper.Map<BlogComment>(blogCommentDTO);
-            blogComment = await blogCommentService.Insert(blogComment);
-            blogCommentDTO = mapper.Map<BlogCommentDTO>(blogComment);
-            return Ok(new { blogComment });
+            int userId = GetCurrentUserId();
+
+            BlogComment comment = new BlogComment();
+            comment.BlogId = blogCommentDTO.BlogId;
+            comment.UserId = userId;
+            comment.Content = blogCommentDTO.Content;
+            comment = await blogCommentService.Insert(comment);
+            blogCommentDTO = mapper.Map<BlogCommentDTO>(comment);
+
+            // Get comment user
+            var user = await userService.GetById(blogCommentDTO.UserId);
+            blogCommentDTO.User = mapper.Map<UserCommentDTO>(user);
+            blogCommentDTO.User.Picture = await component.GetImageAsync(blogCommentDTO.User.PictureId);
+            return Ok(blogCommentDTO);
         }
 
-        // PUT api/<BlogCommentController>/5
+        // PUT: api/<BlogCommentController>/{id}
         [HttpPut("{id}")]
         public async Task<ActionResult<BlogCommentDTO>> Put(int id, [FromBody] BlogCommentDTO blogCommentDTO)
         {
@@ -64,20 +92,36 @@ namespace AcademyProject.Controllers
             {
                 return NotFound();
             }
-            blogComment.BlogId = blogCommentDTO.BlogId;
-            blogComment.UserId = blogCommentDTO.UserId;
+
+            int userId = GetCurrentUserId();
+            if (userId != blogComment.UserId && !User.IsInRole("Administrators") && !User.IsInRole("Moderators"))
+            {
+                return Unauthorized();
+            }
+
             blogComment.Content = blogCommentDTO.Content;
-            blogComment.CreatedAt = blogCommentDTO.CreatedAt;
+            blogComment.UpdatedAt = DateTime.Now;
+            await blogCommentService.Update(blogComment);
 
-            blogCommentDTO = mapper.Map<BlogCommentDTO>(blogComment);
-
-            return Ok(new { blogCommentDTO });
+            blogCommentDTO.UpdatedAt = blogComment.UpdatedAt;
+            return Ok(blogCommentDTO);
         }
 
-        // DELETE api/<BlogCommentController>/5
+        // DELETE: api/<BlogCommentController>/{id}
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var blogComment = await blogCommentService.GetById(id);
+            if (blogComment == null)
+            {
+                return NotFound();
+            }
+            int userId = GetCurrentUserId();
+            if (userId != blogComment.UserId && !User.IsInRole("Administrators") && !User.IsInRole("Moderators"))
+            {
+                return Unauthorized();
+            }
+
             await blogCommentService.Delete(id);
             return Ok();
         }
